Refactor: AvoidFullSQLRequest to extend AbstractSQLPatternCheck for improved SQL pattern handling
